+++ docs/source/conf.py
@@ -1,37 +1,22 @@
 # Configuration file for the Sphinx documentation builder.
 #
-# This file only contains a selection of the most common options. For a full
-# list see the documentation:
+# For the full list of built-in configuration values, see the documentation:
 # https://www.sphinx-doc.org/en/master/usage/configuration.html
 
-# -- Path setup --------------------------------------------------------------
-
-# If extensions (or modules to document with autodoc) are in another directory,
-# add these directories to sys.path here. If the directory is relative to the
-# documentation root, use os.path.abspath to make it absolute, like shown here.
-import os
-import sys
-from importlib import import_module
-from pathlib import Path
-
-src_dir = Path(__file__).resolve().parents[2].joinpath("src")
-sys.path.insert(0, os.fsdecode(src_dir))
-gethash = import_module("gethash")
+import gethash as pkg
 
 # -- Project information -----------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
 project = "Gethash"
 copyright = "2020-2023, xymy"
 author = "xymy"
-
-version = gethash.__version__
-release = gethash.__version__
+version = pkg.__version__
+release = pkg.__version__
 
 # -- General configuration ---------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
-# Add any Sphinx extension module names here, as strings. They can be
-# extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
-# ones.
 extensions = [
     "sphinx.ext.autodoc",
     "sphinx.ext.intersphinx",
@@ -40,44 +25,39 @@ extensions = [
     "myst_parser",
     "sphinx_copybutton",
 ]
-
-# Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]
 
-# List of patterns, relative to source directory, that match files and
-# directories to ignore when looking for source files.
-# This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
-
 # -- Options for HTML output -------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
-# The theme to use for HTML and HTML Help pages.  See the documentation for
-# a list of builtin themes.
 html_theme = "furo"
-
-# Add any paths that contain custom static files (such as style sheets) here,
-# relative to this directory. They are copied after the builtin static files,
-# so a file named "default.css" will overwrite the builtin "default.css".
+html_title = f"{project} {release}"
 html_static_path = ["_static"]
 
-# -- Options for autodoc -------------------------------------------------
+# -- Options for autodoc -----------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/extensions/autodoc.html#module-sphinx.ext.autodoc
 
-autodoc_member_order = "bysource"
+autodoc_default_options = {
+    "members": True,
+    "member-order": "bysource",
+    "show-inheritance": True,
+}
 
 # -- Options for intersphinx -------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/extensions/intersphinx.html#module-sphinx.ext.intersphinx
 
-intersphinx_mapping = {"python": ("https://docs.python.org/3", None)}
+intersphinx_mapping = {
+    "python": ("https://docs.python.org/3/", None),
+}
 
 # -- Options for myst-parser -------------------------------------------------
+# https://myst-parser.readthedocs.io/en/latest/
 
-myst_enable_extensions = [
-    "deflist",
-    "dollarmath",
-    "linkify",
-    "tasklist",
-]
+myst_enable_extensions = ["deflist", "tasklist"]
+myst_heading_anchors = 3
 
 # -- Options for sphinx-copybutton -------------------------------------------
+# https://sphinx-copybutton.readthedocs.io/en/latest/
 
 copybutton_prompt_text = r">>> |\.\.\. |\$ |In \[\d*\]: | {2,5}\.\.\.: | {5,8}: "
 copybutton_prompt_is_regexp = True